Refactor the internal context used for labels and icons
